repomap v1.3.0: add distro field and rhui is string
Introducing possibility to map repositories of various RHEL-like
distributions. Originally it has been expected to cover just RHEL
repositories so the field has not been needed - till now.

All repositories will have specified the distro field to distinguish
repoids of various distributions. We have considered to use the `enum`
type for the distro field, but we discovered this problematic from
`rhui` experience. So both these fields are changed to strings to
make it easier to cover new distributions and clouds in future.

Bumping the version to 1.3.0.

Jira: RHEL-80336